tts_module/requirements.txt
aiosqlite==0.19.0
alembic==1.7.6.dev0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
audioread==3.0.1
Babel==2.13.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.4
blis==0.7.11
Brlapi==0.8.3
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.1
click==8.1.7
cloudpathlib==0.16.0
cmdstanpy==1.2.0
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
confection==0.1.3
contourpy==1.1.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.8
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
exceptiongroup==1.1.3
executing==2.0.1
Faker==19.12.1
fastapi==0.104.1
fasteners==0.14.1
fastjsonschema==2.18.1
filelock==3.13.1
fonttools==4.43.1
fqdn==1.5.1
frozendict==2.3.8
fsspec==2023.10.0
future==0.18.2
greenlet==1.1.2
h11==0.13.0
holidays==0.35
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
importlib-resources==6.1.0
inflect==7.0.0
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.8.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==23.5.0
kiwisolver==1.4.5
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
librosa==0.9.1
llvmlite==0.41.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.7
multitasking==0.0.11
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
noisereduce==3.0.0
notebook==6.4.12
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
patsy==0.5.3
peewee==3.17.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.11.0
pooch==1.8.0
preshed==3.0.9
prometheus-client==0.18.0
prompt-toolkit==3.0.39
prophet==1.1.5
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
referencing==0.30.2
reportlab==3.6.8
requests==2.31.0
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
scikit-learn==1.3.2
scipy==1.11.3
screen-resolution-extra==0.0.0
seaborn==0.13.0
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post10
smart-open==6.4.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
sox==1.4.1
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.22
srsly==2.4.8
stack-data==0.6.3
stanio==0.3.0
starlette==0.27.0
statsmodels==0.14.0
sympy==1.12
systemd-python==234
tensorboardX==2.6.2.2
terminado==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.0
torchaudio==2.1.0
torchdata==0.7.0
torchtext==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.7
uri-template==1.3.0
urllib3==2.0.7
usb-creator==0.3.7
uvicorn==0.15.0
virtualenv==20.24.6
wadllib==1.3.6
wasabi==1.1.2
wcwidth==0.2.9
weasel==0.3.3
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
wsproto==1.0.0
xdg==5
xkit==0.0.0
yfinance==0.2.31
zipp==1.0.0
